Move tasty.CompilerInterface to scala.internal
This class is not part of the user interface.
